requirements.txt
# 金融客服智能体项目依赖

# LangChain核心框架
langchain>=0.3.0
langchain-community>=0.3.0
langchain-mcp-adapters>=0.2.0
langchain-core>=1.2.0

# 向量数据库
chromadb>=0.5.0
qdrant-client>=1.12.0

# LLM和嵌入模型
openai>=1.0.0
dashscope>=1.0.0

# 配置管理
python-dotenv>=1.0.0

# MCP协议支持
mcp>=1.26.0

# HTTP客户端
httpx>=0.28.0
